Fix declaration mismatch

+++ include/pgm_index.hpp
@@ -58,7 +58,7 @@ struct ApproxPos {
 template<typename K, size_t Epsilon = 64, size_t EpsilonRecursive = 4, typename Floating = double>
 class PGMIndex {
 protected:
-    template<typename, size_t, size_t, typename>
+    template<typename, size_t, uint8_t, typename>
     friend class InMemoryPGMIndex;
 
     static_assert(Epsilon > 0);
